Renaming to services folder
