+++ SplunkAppForWazuh/appserver/static/js/directives/agentsTable.js
