+++ SplunkAppForWazuh/appserver/static/js/directives/toaster.js
