+++ SplunkAppForWazuh/appserver/static/js/services/apiService.js
