+++ SplunkAppForWazuh/appserver/static/js/services/credentialService.js
