+++ SplunkAppForWazuh/appserver/static/js/services/indexService.js
