+++ SplunkAppForWazuh/appserver/static/js/services/localStorage.js
