+++ SplunkAppForWazuh/appserver/static/js/services/promisedReq.js
