+++ SplunkAppForWazuh/appserver/static/js/viewControllers/APIconf.js
@@ -19,9 +19,9 @@ require([
   "jquery",
   "splunkjs/mvc/layoutview",
   "splunkjs/mvc/simplexml/dashboardview",
-  "/static/app/SplunkAppForWazuh/js/utilLib/credentialService.js",
-  "/static/app/SplunkAppForWazuh/js/utilLib/indexService.js",
-  "/static/app/SplunkAppForWazuh/js/customViews/toaster.js",
+  "/static/app/SplunkAppForWazuh/js/services/credentialService.js",
+  "/static/app/SplunkAppForWazuh/js/services/indexService.js",
+  "/static/app/SplunkAppForWazuh/js/directives/toaster.js",
   "splunkjs/mvc/simplexml/searcheventhandler",
   "splunkjs/mvc/simpleform/input/dropdown",
   "splunkjs/mvc/searchmanager",
